pattern tween fix

+++ tween.py
@@ -174,15 +174,15 @@ class PatternTween(Tween):
         for i, cell in enumerate(pattern):
             if i % start_div == 0 and i % target_div == 0:
                 if random.random() > position:
-                    pattern[i] = start_steps[i / start_div]
+                    pattern[i] = start_steps[int(i / start_div)]
                 else:
-                    pattern[i] = target_steps[i / target_div]
+                    pattern[i] = target_steps[int(i / target_div)]
             elif i % start_div == 0:     
                 if random.random() > position:               
-                    pattern[i] = start_steps[i / start_div]
+                    pattern[i] = start_steps[int(i / start_div)]
             elif i % target_div == 0:
                 if random.random() <= position:
-                    pattern[i] = target_steps[i / target_div]
+                    pattern[i] = target_steps[int(i / target_div)]
         pattern = Pattern(pattern)
         return pattern
